Implement Exploratory Daemon (#19)
Implement Exploratory Daemon.

See cmudb/exploration/README.md for more details.

This commit adds the functionality to take online copies of a
postgres database. The goal of the copy is to be able to run fast
but not necessarily have the most up to date data. The process of
taking a copy is the following:

1. Run checkpoint on original database (This is not strictly
required).
2. Take a ZFS snapshot and clone of pgdata directory.
3. Run pg_resetwal on ZFS copy. This will prune the log to some
consistent point which makes starting the new instance much faster.
However it also results in some lost data.
4. Start exploratory docker container which uses the copied pgdata
directory to start a postgres instance.

The process of tearing down an exploratory instance is the
following:

1. Stop postgres instance.
2. Shutdown docker container.
3. Delete copied data.

The daemon needs to be run on a system with an existing postgres
instance and ZFS installed.

The current implementation just takes a single copy, runs some
random queries, and tears down the copy. In the future this
functionality will be daemonized to allow on demand copies to
be made.

+++ cmudb/env/docker-entrypoint.sh
@@ -25,6 +25,20 @@ PGPORT=${PGPORT}    # The port to listen on.
 NP_REPLICATION_TYPE=${NP_REPLICATION_TYPE}          # Must be "primary" or "replica"
 NP_REPLICATION_USER=${NP_REPLICATION_USER}          # Replication user.
 NP_REPLICATION_PASSWORD=${NP_REPLICATION_PASSWORD}  # Replication password.
+NP_WAL_LEVEL=${NP_WAL_LEVEL}                        # The wal_level setting in PostgreSQL.
+
+# The primary must have these defined.
+NP_REPLICATION_PHYSICAL_SLOTS=${NP_REPLICATION_PHYSICAL_SLOTS}
+
+# These settings need to be defined on ALL replicas.
+NP_PRIMARY_NAME=${NP_PRIMARY_NAME}                  # The name of the primary on the Docker instance.
+NP_PRIMARY_USERNAME=${NP_PRIMARY_USERNAME}          # The username to connect to the primary.
+NP_PRIMARY_PORT=${NP_PRIMARY_PORT}                  # The port on the primary.
+
+# Physical replicas.
+NP_REPLICATION_PHYSICAL_SLOT=${NP_REPLICATION_PHYSICAL_SLOT}    # The physical slot that this replica will use.
+
+
 
 # =====================================================================
 # Default environment variable values.
@@ -46,6 +60,10 @@ if [ -z "$PGPORT" ]; then
   PGPORT=15721
 fi
 
+if [ -z "$NP_WAL_LEVEL" ]; then
+  NP_WAL_LEVEL="replica"
+fi
+
 # =====================================================================
 # Helper functions.
 # =====================================================================
@@ -55,11 +73,28 @@ _pgctl_start() {
 }
 
 _pg_stop() {
-  ${BIN_DIR}/pg_ctl --pgdata=${PGDATA} -w stop
+  if [ "${NP_REPLICATION_TYPE}" = "exploratory" ]; then
+    ${BIN_DIR}/pg_ctl --pgdata=${PGDATA} -w --mode=immediate stop
+  else
+    ${BIN_DIR}/pg_ctl --pgdata=${PGDATA} -w stop
+  fi
 }
 
 _pg_start() {
-  ${BIN_DIR}/postgres "-D" "${PGDATA}" -p 15721
+  if [ "${NP_REPLICATION_TYPE}" = "primary" ]; then
+      (
+        sleep 5
+        IFS=','
+        for publication in $(echo "${NP_REPLICATION_PUBLICATION_NAMES[@]}"); do
+          echo $publication
+          # Create logical replication publication(s).
+          # Note that for updates/deletes to work, each table may need REPLICA IDENTITY to be manually set.
+          ${BIN_DIR}/psql -c "create publication $publication for all tables" postgres
+        done
+      ) &
+    fi
+
+    ${BIN_DIR}/postgres "-D" "${PGDATA}" -p ${PGPORT}
 }
 
 _pg_initdb() {
@@ -87,6 +122,8 @@ _pg_create_user_and_db() {
   ${BIN_DIR}/psql -c "create database ${POSTGRES_DB} with owner = '${POSTGRES_USER}'" postgres
   # Enable monitoring for the created user.
   ${BIN_DIR}/psql -c "grant pg_monitor to ${POSTGRES_USER}" postgres
+  # Make the created user a superuser.
+  ${BIN_DIR}/psql -c "alter user ${POSTGRES_USER} with superuser" postgres
 }
 
 _pg_setup_replication() {
@@ -96,7 +133,7 @@ _pg_setup_replication() {
   # See PostgreSQL docs for complete description of parameters.
 
   # wal_level: How much information to ship over.
-  echo "wal_level = replica" >> ${AUTO_CONF}
+  echo "wal_level = ${NP_WAL_LEVEL}" >> ${AUTO_CONF}
   # hot_standby: True to enable connecting and running queries during recovery.
   echo "hot_standby = on" >> ${AUTO_CONF}
   # max_wal_senders: Maximum number of concurrent connections to standby/backup clients.
@@ -105,6 +142,32 @@ _pg_setup_replication() {
   echo "max_replication_slots = 10" >> ${AUTO_CONF}
   # hot_standby_feedback: True if standby should tell primary about what queries are currently executing.
   echo "hot_standby_feedback = on" >> ${AUTO_CONF}
+
+  # PGTune configs (Most configs divided by two to accommodate two instances)
+  # DB Version: 13
+    # OS Type: linux
+    # DB Type: oltp
+    # Total Memory (RAM): 188 GB
+    # CPUs num: 80
+    # Number of connections: 50
+    # Data Storage: SSD
+
+  echo "max_connections = 100" >> ${AUTO_CONF}
+  echo "shared_buffers = 23.5GB" >> ${AUTO_CONF}
+  echo "effective_cache_size = 70.5GB" >> ${AUTO_CONF}
+  echo "maintenance_work_mem = 2GB" >> ${AUTO_CONF}
+  echo "checkpoint_completion_target = 0.9" >> ${AUTO_CONF}
+  echo "wal_buffers = 16MB" >> ${AUTO_CONF}
+  echo "default_statistics_target = 100" >> ${AUTO_CONF}
+  echo "random_page_cost = 1.1" >> ${AUTO_CONF}
+  echo "effective_io_concurrency = 200" >> ${AUTO_CONF}
+  echo "work_mem = 123207.5kB" >> ${AUTO_CONF}
+  echo "min_wal_size = 2GB" >> ${AUTO_CONF}
+  echo "max_wal_size = 8GB" >> ${AUTO_CONF}
+  echo "max_worker_processes = 80" >> ${AUTO_CONF}
+  echo "max_parallel_workers_per_gather = 4" >> ${AUTO_CONF}
+  echo "max_parallel_workers = 80" >> ${AUTO_CONF}
+  echo "max_parallel_maintenance_workers = 4" >> ${AUTO_CONF}
 
   if [ "${NP_REPLICATION_TYPE}" = "primary" ]; then
     # ===============================
@@ -117,20 +180,31 @@ _pg_setup_replication() {
     echo "host replication ${NP_REPLICATION_USER} 0.0.0.0/0 md5" >> ${HBA_CONF}
     # Reload configuration.
     ${BIN_DIR}/psql -c "select pg_reload_conf()" postgres
-    # Create replication slot for replica.
-    ${BIN_DIR}/psql -c "select pg_create_physical_replication_slot('replication_slot_replica1')" postgres
+    echo ${NP_REPLICATION_PHYSICAL_SLOTS}
+    (
+      IFS=','
+      for slot in $(echo "${NP_REPLICATION_PHYSICAL_SLOTS[@]}"); do
+        echo $slot
+        # Create replication slot(s) for replica.
+        ${BIN_DIR}/psql -c "select pg_create_physical_replication_slot('$slot')" postgres
+      done
+    )
   fi
 }
 
 # All the steps required to start up PostgreSQL.
 _pg_start_all() {
-  _pg_initdb              # Initialize a new PostgreSQL cluster.
-  _pg_config              # Write any configuration options required.
-  _pgctl_start            # Start the PostgreSQL cluster.
-  _pg_create_user_and_db  # Create the specified user and database.
+  if [ -e "${PGDATA}/base" ]; then
+    _pgctl_start
+  else
+    _pg_initdb              # Initialize a new PostgreSQL cluster.
+    _pg_config              # Write any configuration options required.
+    _pgctl_start            # Start the PostgreSQL cluster.
+    _pg_create_user_and_db  # Create the specified user and database.
 
-  if [ ! -z "${NP_REPLICATION_TYPE}" ]; then
-    _pg_setup_replication
+    if [ ! -z "${NP_REPLICATION_TYPE}" ]; then
+      _pg_setup_replication
+    fi
   fi
 }
 
@@ -138,29 +212,57 @@ _pg_start_all() {
 # Main logic.
 # =====================================================================
 
-main() {
-  # Only initdb if this is not a replica. The replica will recover from backup.
-  if [ ! "${NP_REPLICATION_TYPE}" = "replica" ]; then
-    # This is a single-node or the primary.
-    _pg_start_all
-    _pg_stop
-    _pg_start
-  else
-    # This is a replica.
-    while true ; do
-      # TODO(WAN): Issue #6 Note that there is a potential race here where the primary restarts and healthcheck succeeds.
-      sleep 10
-      ${BIN_DIR}/pg_isready --host primary --port 15721 --username noisepage
-      READY_CHECK=$?
-      if [ "$READY_CHECK" = "0" ]; then
-        break
-      fi
-    done
+_wait_for_primary() {
+  while true ; do
+    # TODO(WAN): Issue #6 Note that there is a potential race here where the primary restarts and healthcheck succeeds.
+    sleep 10
+    ${BIN_DIR}/pg_isready --host=${NP_PRIMARY_NAME} --port=${NP_PRIMARY_PORT} --username=${NP_PRIMARY_USERNAME}
+    READY_CHECK=$?
+    if [ "$READY_CHECK" = "0" ]; then
+      break
+    fi
+  done
+}
 
-    rm -rf ${PGDATA}/*
-    # Initialize replica backup from primary.
-    echo passyMcPassword | ${BIN_DIR}/pg_basebackup --host primary --username replicator --port 15721 --pgdata=${PGDATA} --format=p --wal-method=stream --progress --write-recovery-conf --slot replication_slot_replica1
-    _pg_start
+_main_primary() {
+  _pg_start_all
+  _pg_stop
+  _pg_start
+}
+
+_main_replica() {
+  _wait_for_primary
+
+  # Initialize replica backup from primary.
+  rm -rf "${PGDATA:?}/"*
+  echo ${NP_REPLICATION_PASSWORD} | ${BIN_DIR}/pg_basebackup --host ${NP_PRIMARY_NAME} --username ${NP_REPLICATION_USER} --port ${NP_PRIMARY_PORT} --pgdata=${PGDATA} --format=p --wal-method=stream --progress --write-recovery-conf --slot ${NP_REPLICATION_PHYSICAL_SLOT}
+  _pg_start
+}
+
+_main_exploratory() {
+  sudo chmod 700 "${PGDATA}"
+  sudo chown -R terrier:terrier "${PGDATA}"
+  rm -f "${PGDATA}/postmaster.pid"
+  rm -f "${PGDATA}/standby.signal"
+  "${BIN_DIR}"/pg_resetwal -f "${PGDATA}"
+  _pg_start
+}
+
+_cleanup() {
+  _pg_stop
+}
+
+main() {
+  trap 'cleanup' SIGTERM
+  if [ -z "${NP_REPLICATION_TYPE}" ] || [ "${NP_REPLICATION_TYPE}" = "primary" ]; then
+    _main_primary
+  elif [ "${NP_REPLICATION_TYPE}" = "replica" ]; then
+    _main_replica
+  elif [ "${NP_REPLICATION_TYPE}" = "exploratory" ]; then
+    _main_exploratory
+  else
+    echo "Unknown replication type: ${NP_REPLICATION_TYPE}"
+    exit 1
   fi
 }
